requirements.txt
Django>=4.0
channels>=4.0
channels-redis>=4.0
djangorestframework>=3.13
djangorestframework-simplejwt>=5.2
asgiref>=3.5
python-chess>=1.999
redis>=4.0
mysqlclient>=2.0
daphne>=4.0
pytest>=7.0
pytest-django>=4.0